+++ VSDocumentReopen/VSDocumentReopen/ReopenPackage.cs
@@ -9,6 +9,7 @@ using EnvDTE;
 using EnvDTE80;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
+using VSDocumentReopen.Commands;
 using Task = System.Threading.Tasks.Task;
 
 namespace VSDocumentReopen
