+++ VSDocumentReopen/VSDocumentReopen/VSDocumentReopen.csproj
@@ -56,8 +56,8 @@
   <ItemGroup>
     <Compile Include="DocumentTracker.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ClearHistory.cs" />
-    <Compile Include="Reopen.cs" />
+    <Compile Include="Commands\ClearHistory.cs" />
+    <Compile Include="Commands\Reopen.cs" />
     <Compile Include="ReopenPackage.cs" />
   </ItemGroup>
   <ItemGroup>
